Extract coordinate arrays from ROMS.

Returns a RasterStack of the given variable names.

romscoords(x, spatial = c("lon_u", "lat_u"), ncdf = TRUE,
  transpose = FALSE, ...)

Arguments

x
ROMS file name

spatial
names of coordinate variables (e.g. lon_u, lat_u)

ncdf
default to NetCDF no matter what file name

transpose
the extents (ROMS is FALSE, Access is TRUE)

...
unused

Value

RasterStack with two layers of the 2D-variables

Details

The two layers from the model output are used to define the real-world space. This is used to create a boundary romsboundary, to map real-world objects into grid space romscoords and to generate graticules for mapping into the grid space with graphics::contour.

Examples

# NOT RUN {
  coord <- romscoord("roms.nc")
# }
## with in-built fake data
plot(ice_fake, asp = 0.5)

#> Error in as.double(y): cannot coerce type 'S4' to vector of type 'double'

contour(ice_coords[[1]], add = TRUE, levels = seq(-165, 165, by = 15))

#> Error in contour.default(x = xFromCol(x, 1:ncol(x)), y = yFromRow(x, nrow(x):1),     z = t((getValues(x, format = "matrix"))[nrow(x):1, ]), ...): plot.new has not been called yet

contour(ice_coords[[2]], add = TRUE)

#> Error in contour.default(x = xFromCol(x, 1:ncol(x)), y = yFromRow(x, nrow(x):1),     z = t((getValues(x, format = "matrix"))[nrow(x):1, ]), ...): plot.new has not been called yet