ROMS single slice 2D layer Extract a data layer from ROMS by name and 4-D slice.

romsdata always works in the first two dimensions (x-y), the more specialist functions will work in the space indicated by their name roms_xy, roms_xt and so on.

roms_xy(x, varname, slice = c(1L, 1L), transpose = TRUE, ...)

roms_xz(x, varname, slice = c(1L, 1L), transpose = TRUE, ...)

roms_xt(x, varname, slice = c(1L, 1L), transpose = TRUE, ...)

roms_yz(x, varname, slice = c(1L, 1L), transpose = TRUE, ...)

roms_yt(x, varname, slice = c(1L, 1L), transpose = TRUE, ...)

roms_zt(x, varname, slice = c(1L, 1L), transpose = TRUE, ...)

romsdata(x, varname, slice = c(1L, 1L), ncdf = TRUE, transpose = TRUE,
  ...)

Arguments

x
ROMS file name

varname
name of ROMS variable

slice
index in w and t (depth and time), defaults to first encountered

transpose
the extents (ROMS is FALSE, Access is TRUE)

...
unused

ncdf
default to TRUE, set to FALSE to allow raster format detection brick

Value

RasterLayer

Examples

#x <- raadtools:::cpolarfiles()$fullname[1]
#plot(roms_xy(x, "u"))
#plot(roms_xz(x, "u", slice = c(392L,1L)), asp = NA)
#plot(roms_xt(x, "u", slice = c(392L,1L)), asp = NA)

#plot(roms_yz(x, "u"))
#plot(roms_yt(x, "u", slice = c(1L,1L)), asp = NA)
#plot(roms_zt(x, "u", slice = c(1L, 392L)), asp = NA)